feat(lex-store,lex-api): publish_program helper; HTTP /v1/publish via apply_operation
- Store::publish_program is the shared entry point: build old views,
  diff_to_ops, loop apply_operation. Used by both cmd_publish (CLI)
  and publish_handler (HTTP).
- compute_diff and its helpers moved from lex-cli to lex-vcs so both
  the CLI and the HTTP API can compute a DiffReport without a layering
  violation (lex-cli -> lex-api is a hard cycle).
- HTTP /v1/publish returns { ops: [...], head_op: ... } matching the
  CLI's JSON envelope.
- Tests updated to extract stage_id from ops[0].kind.stage_id.

Refs #129.
