+++ crates/lex-store/src/lib.rs
@@ -31,7 +31,7 @@ mod branches;
 
 pub use lex_vcs::{OpId, Operation, OperationKind, OperationRecord, StageTransition};
 pub use model::{Lifecycle, Metadata, Spec, StageStatus, Test, Transition};
-pub use store::{StageHistoryEntry, Store, StoreError};
+pub use store::{PublishOp, PublishOutcome, StageHistoryEntry, Store, StoreError};
 pub use branches::{
     Branch, MergeConflict, MergeEntry, MergeRecord, MergeReport, MergeSummary,
     DEFAULT_BRANCH,
