+++ crates/lex-vcs/src/lib.rs
@@ -30,6 +30,7 @@
 
 mod apply;
 mod canonical;
+mod compute_diff;
 pub mod diff_report;
 mod diff_to_ops;
 mod merge;
@@ -37,6 +38,7 @@ mod op_log;
 mod operation;
 
 pub use apply::{apply, ApplyError, NewHead};
+pub use compute_diff::{compute_diff, effect_label, render_signature};
 pub use diff_report::DiffReport;
 pub use diff_to_ops::{diff_to_ops, DiffInputs, DiffMappingError, ImportMap};
 pub use merge::{merge, ConflictKind, MergeOutcome, MergeOutput};
